+++ src/main/java/heronarts/lx/command/LXCommand.java
@@ -1759,177 +1759,7 @@ public abstract class LXCommand {
 
   public static class Snapshots {
 
-    public static class AddSnapshot extends LXCommand {
-
-      private ComponentReference<LXSnapshot> snapshot;
-      private JsonObject initialObj = null;
-      private JsonObject snapshotObj = null;
-      private final int index;
-
-      public AddSnapshot() {
-        this.index = -1;
-      }
-
-      public AddSnapshot(JsonObject snapshotObj, int index) {
-        this.initialObj = snapshotObj;
-        this.index = index;
-      }
-
-      @Override
-      public String getDescription() {
-        return "Add Snapshot";
-      }
-
-      @Override
-      public void perform(LX lx) {
-        if (this.snapshotObj == null) {
-          LXSnapshot instance;
-          if (this.initialObj != null) {
-            instance = new LXSnapshot(lx);
-            instance.load(lx, this.initialObj);
-            lx.engine.snapshots.addSnapshot(instance, this.index);
-          } else {
-            instance = lx.engine.snapshots.addSnapshot();
-          }
-          this.snapshot = new ComponentReference<LXSnapshot>(instance);
-          this.snapshotObj = LXSerializable.Utils.toObject(lx, instance);
-        } else {
-          LXSnapshot instance = new LXSnapshot(lx);
-          instance.load(lx, this.snapshotObj);
-          this.snapshot = new ComponentReference<LXSnapshot>(instance);
-          lx.engine.snapshots.addSnapshot(instance);
-        }
-      }
-
-      @Override
-      public void undo(LX lx) {
-        lx.engine.snapshots.removeSnapshot(this.snapshot.get());
-      }
-    }
-
-    public static class MoveSnapshot extends LXCommand {
-
-      private final ComponentReference<LXSnapshot> snapshot;
-      private final int fromIndex;
-      private final int toIndex;
-
-      public MoveSnapshot(LXSnapshot snapshot, int toIndex) {
-        this.snapshot = new ComponentReference<LXSnapshot>(snapshot);
-        this.fromIndex = snapshot.getIndex();
-        this.toIndex = toIndex;
-      }
-
-      @Override
-      public String getDescription() {
-        return "Move Snapshot";
-      }
-
-      @Override
-      public void perform(LX lx) {
-        lx.engine.snapshots.moveSnapshot(this.snapshot.get(), this.toIndex);
-      }
-
-      @Override
-      public void undo(LX lx) {
-        lx.engine.snapshots.moveSnapshot(this.snapshot.get(), this.fromIndex);
-      }
-    }
-
-    public static class RemoveSnapshot extends RemoveComponent {
-
-      private final ComponentReference<LXSnapshot> snapshot;
-      private final JsonObject snapshotObj;
-      private final int snapshotIndex;
-
-      public RemoveSnapshot(LXSnapshot snapshot) {
-        super(snapshot);
-        this.snapshot = new ComponentReference<LXSnapshot>(snapshot);
-        this.snapshotObj = LXSerializable.Utils.toObject(snapshot);
-        this.snapshotIndex = snapshot.getIndex();
-      }
-
-      @Override
-      public String getDescription() {
-        return "Delete Snapshot";
-      }
-
-      @Override
-      public void perform(LX lx) {
-        lx.engine.snapshots.removeSnapshot(this.snapshot.get());
-      }
-
-      @Override
-      public void undo(LX lx) throws InvalidCommandException {
-        LXSnapshot snapshot = new LXSnapshot(lx);
-        snapshot.load(lx, this.snapshotObj);
-        lx.engine.snapshots.addSnapshot(snapshot, this.snapshotIndex);
-        super.undo(lx);
-      }
-    }
-
-    public static class Update extends LXCommand {
-
-      private final ComponentReference<LXSnapshot> snapshot;
-      private JsonObject previousState;
-
-      public Update(LXSnapshot snapshot) {
-        this.snapshot = new ComponentReference<LXSnapshot>(snapshot);
-      }
-
-      @Override
-      public String getDescription() {
-        return "Update Snapshot";
-      }
-
-      @Override
-      public void perform(LX lx) throws InvalidCommandException {
-        this.previousState = LXSerializable.Utils.toObject(lx, this.snapshot.get());
-        this.snapshot.get().update();
-      }
-
-      @Override
-      public void undo(LX lx) throws InvalidCommandException {
-        this.snapshot.get().load(lx, this.previousState);
-      }
-
-    }
-
-    public static class Recall extends LXCommand {
-
-      private final ComponentReference<LXSnapshot> snapshot;
-      private final List<LXCommand> commands = new ArrayList<LXCommand>();
-      private boolean recalled = false;
-
-      public Recall(LXSnapshot snapshot) {
-        this.snapshot = new ComponentReference<LXSnapshot>(snapshot);
-      }
-
-      @Override
-      public void perform(LX lx) {
-        this.commands.clear();
-        this.recalled = lx.engine.snapshots.recall(this.snapshot.get(), this.commands);
-      }
-
-      @Override
-      public void undo(LX lx) throws InvalidCommandException {
-        for (LXCommand command : this.commands) {
-          command.undo(lx);
-        }
-      }
-
-      @Override
-      public boolean isIgnored() {
-        return !this.recalled;
-      }
-
-      @Override
-      public String getDescription() {
-        return "Recall Snapshot";
-      }
-    }
-
-    private static class RemoveView extends LXCommand {
-
+     private static class RemoveView extends LXCommand {
       private ComponentReference<LXSnapshot> snapshot;
       private LXSnapshot.View view;
       private final JsonObject viewObj;
